@@ -3,22 +3,27 @@ source 'https://rubygems.org'
 branch = ENV.fetch('SOLIDUS_BRANCH', 'master')
 gem 'solidus', github: 'solidusio/solidus', branch: branch
 
-# Needed to help Bundler figure out how to resolve dependencies, otherwise it takes forever to
-# resolve them
+# Needed to help Bundler figure out how to resolve dependencies,
+# otherwise it takes forever to resolve them
 if branch == 'master' || Gem::Version.new(branch[1..-1]) >= Gem::Version.new('2.10.0')
   gem 'rails', '~> 6.0'
 else
-  gem 'rails', '~> 5.0'
+  gem 'rails', '~> 5.2'
 end
 
 group :test do
   gem 'rails-controller-testing'
 end
 
-gem 'rake', '< 11.0'
-gem 'capybara-screenshot', group: :test
+case ENV['DB']
+when 'mysql'
+  gem 'mysql2'
+when 'postgresql'
+  gem 'pg'
+else
+  gem 'sqlite3'
+end
 
-gem 'pg'
-gem 'mysql2'
+gem 'solidus_extension_dev_tools', github: 'solidusio-contrib/solidus_extension_dev_tools'
 
 gemspec
