@@ -1,27 +1,13 @@
 # frozen_string_literal: true
 
-require "simplecov"
-SimpleCov.start "rails"
+ENV['RAILS_ENV'] ||= 'test'
 
-ENV["RAILS_ENV"] ||= "test"
+require 'solidus_extension_dev_tools/rspec/coverage'
 
 require File.expand_path('dummy/config/environment.rb', __dir__)
 
-require "solidus_support"
-require "solidus_support/extension/feature_helper"
-require 'spree/testing_support/controller_requests'
-require 'spree/testing_support/capybara_ext'
+require 'solidus_extension_dev_tools/rspec/feature_helper'
 
-Dir[File.join(File.dirname(__FILE__), "support/**/*.rb")].each { |f| require f }
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
 
-FactoryBot.find_definitions
-
-RSpec.configure do |config|
-  config.infer_spec_type_from_file_location!
-  config.raise_errors_for_deprecations!
-
-  config.example_status_persistence_file_path = "./spec/examples.txt"
-
-  config.include Spree::TestingSupport::UrlHelpers
-  config.include Spree::TestingSupport::ControllerRequests, type: :controller
-end
+require 'spree_handling_fees/factories'
